% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualoc.df-data.R
\docType{data}
\name{manualoc.df}
\alias{manualoc.df}
\title{Data frame of manualoc() selections.}
\format{A data frame with 11 rows and 6 variables: \describe{ 
 \item{sound.files}{recording names}
 \item{selec}{selection numbers within recording}
 \item{start}{start times of selected call or element}
 \item{end}{end times of selected call or element}
 \item{sel.comment}{selection comments}
 \item{rec.comment}{recording comments}
}}
\source{
Marcelo Araya Salas, warbleR
}
\usage{
data(manualoc.df)
}
\description{
A data frame containing information for calls selected using \code{\link{manualoc}}.
}
\keyword{datasets}

