% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.files-data.R
\docType{data}
\name{selection.files}
\alias{selection.files}
\title{Selections files from Raven and Syrinx.}
\format{\describe{
  
  \item{selection.files}{Selections from the commercial software `Raven` and `Syrinx`}
  
  }}
\usage{
data(selection.files)
}
\description{
Selections files from Raven and Syrinx.
}
\keyword{datasets}

